Error Handling

+++ server/src/ApiControllers/userController.ts
@@ -2,6 +2,7 @@ import { inject, injectable } from "inversify";
 import { Request, Response } from 'express';
 import { User } from '../Models/user';
 import { IUserService } from "../ServiceLayer/userService";
+import { missingParametersError } from "./controllerErrorHelpers";
 
 
 export interface IUserController {
@@ -25,24 +26,39 @@ export class UserController implements IUserController {
         const username = request.body.username;
         const password = request.body.password;
 
+        if(!username || !password) {
+            throw missingParametersError(["username", "password"])
+        }
+
         this.userService.createUser(username, password)
-            .then(result => response.json(result));
+            .then(result => response.json(result))
+            .catch(error => next(error));
     }
 
     deleteUser = (request: Request, response: Response, next: Function): void => {
 
         const userid = request.body.userid;
 
+        if(!userid) {
+            throw missingParametersError(["userid"])
+        }
+
         this.userService.deleteUser(userid)
-            .then(result => response.json(result));
+            .then(result => response.json(result))
+            .catch(error => next(error));
     }
 
     updateUser = (request: Request, response: Response, next: Function): void => {
 
         const user: User = request.body;
 
+        if(!user.userid) {
+            throw missingParametersError(["userid"])
+        }
+
         this.userService.updateUser(user)
-            .then(result => response.json(result));
+            .then(result => response.json(result))
+            .catch(error => next(error));
     }
 
     getUsers = (request: Request, response: Response, next: Function): void => {
@@ -65,6 +81,7 @@ export class UserController implements IUserController {
         }
 
         query
-            .then((users: any) => response.json(users));
+            .then((users: any) => response.json(users))
+            .catch(error => next(error));
     }
 }
